requirements.txt
rasterio
geopandas
numpy
torch
torchvision
segmentation-models-pytorch
pillow
pyproj
shapely
matplotlib
pandas
